bef/agents/specialized/image-generator/requirements.txt
# Image Generator — Requirements

# HTTP requests
requests>=2.31.0

# Image processing
Pillow>=10.2.0

# Progress bar
tqdm>=4.66.1

# Environment variables
python-dotenv>=1.0.0
